+++ src/App/AppEnv.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE MultiParamTypeClasses  #-}
 {-# LANGUAGE TemplateHaskell        #-}
-module App.AppEnv
-where
 
-import Antiope.Env    (Env, HasEnv (..))
+module App.AppEnv where
+
+import Antiope.Env    (Env)
 import App.Options
 import Arbor.Logger   (LogLevel, TimedFastLogger)
 import Control.Lens
@@ -24,30 +24,3 @@ data AppEnv = AppEnv
 
 makeClassy ''AppLogger
 makeClassy ''AppEnv
-
-instance HasEnv AppEnv where
-  environment = appEnvAws
-
-class HasStatsClient a where
-  statsClient :: Lens' a StatsClient
-
-instance HasStatsClient StatsClient where
-  statsClient = id
-
-instance HasStatsClient AppEnv where
-  statsClient = appStatsClient
-
-instance HasKafkaConfig AppEnv where
-  kafkaConfig = appOptions . kafkaConfig
-
-instance HasStatsConfig AppEnv where
-  statsConfig = appOptions . statsConfig
-
-instance HasAppLogger AppEnv where
-  appLogger = appEnv . appLog
-
-instance HasAwsConfig AppEnv where
-  awsConfig = appOptions . optAwsConfig
-
-instance HasStoreConfig AppEnv where
-  storeConfig = appOptions . optStoreConfig
